Refactored modules.py into a collection of modules. Updated dependency in other modules. Updated api in RTD. Updated roadmap for kernel methods.
