+++ docs/api_ref.rst
@@ -12,6 +12,7 @@ Below you'll find the complete public API, grouped by module.
    dymad.data
    dymad.losses
    dymad.models
+   dymad.modules
    dymad.numerics
    dymad.sako
    dymad.training
