+++ src/dymad/modules/__init__.py
@@ -1,0 +1,15 @@
+from dymad.modules.collection import make_autoencoder
+from dymad.modules.gnn import GNN, ResBlockGNN, IdenCatGNN
+from dymad.modules.mlp import MLP, ResBlockMLP, IdenCatMLP
+from dymad.modules.linear import FlexLinear
+
+__all__ = [
+    "FlexLinear",
+    "GNN",
+    "IdenCatGNN",
+    "IdenCatMLP",
+    "make_autoencoder",
+    "MLP",
+    "ResBlockGNN",
+    "ResBlockMLP",
+]
