+++ src/dymad/utils/__init__.py
@@ -1,6 +1,6 @@
 from dymad.utils.checkpoint import load_checkpoint, load_model, save_checkpoint
+from dymad.utils.control import ControlInterpolator
 from dymad.utils.misc import load_config, setup_logging
-from dymad.utils.modules import ControlInterpolator, FlexLinear, GNN, IdenCatGNN, IdenCatMLP, make_autoencoder, MLP, ResBlockGNN, ResBlockMLP
 from dymad.utils.plot import plot_summary, plot_trajectory, plot_hist
 from dymad.utils.prediction import predict_continuous, predict_continuous_exp, predict_discrete, predict_discrete_exp, \
     predict_graph_continuous, predict_graph_discrete
@@ -9,16 +9,10 @@ from dymad.utils.scheduler import make_scheduler
 
 __all__ = [
     "ControlInterpolator",
-    "FlexLinear",
-    "GNN",
-    "IdenCatGNN",
-    "IdenCatMLP",
     "load_checkpoint",
     "load_config",
     "load_model",
-    "make_autoencoder",
     "make_scheduler",
-    "MLP",
     "plot_hist",
     "plot_summary",
     "plot_trajectory",
@@ -28,8 +22,6 @@ __all__ = [
     "predict_discrete_exp",
     "predict_graph_continuous",
     "predict_graph_discrete",
-    "ResBlockGNN",
-    "ResBlockMLP",
     "save_checkpoint",
     "setup_logging",
     "TrajectorySampler",
